move config files for server and next
